+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.749] - 2024-09-30 - implement add-cachedciminstance
+
+## [0.0.748] - 2024-09-30 - implement add-cachedciminstance
+
 ## [0.0.747] - 2024-09-29 - rename task in psakefile
 
 ## [0.0.746] - 2024-09-29 - bugfix missing itempath param when calling resolve-ace